🚧 Separating runner.rs from lib.rs
runner.rs will contain the `Config` struct and `run` function, this used
to be part of `lib.rs`, but was only used by main.rs, and not
recommended for those using this crate as a library.

Now, that functionality will be part of a private module of the binary
crate, and the `file_ops.rs` will basically be the new `lib.rs`. The
`runner` module will be responsible for using the library crate.